requirements.txt
Django==5.2
requests==2.31.0
celery==5.5.2
redis==5.2.1
djangorestframework==3.16.0
django-celery-beat==2.8.0
djangorestframework-simplejwt==5.5.0